chore(api): creating restapi module #24
Made changes to pom.xml files to change project to Java 11 (to support
spring boot). Also added plugins, dependencies and module to pom.xml
files. Generated the restapi module using spring initializer
(https://start.spring.io/) and added the generated files to the module.
Tests for this module currently do not work, but the project runs as
usual.

Breaking changes: If test in module is uncommented, project doesn't build.

In this