feat: API now loads calendar.json from resources
Added module-info.java to api module and renamed it it to restapi /
apiservice to fit better with module naming convention. Also added
calendar.json file to resources folder in api module. This file is
not synced with the rest of the project yet, but shows that the api is
capable of properly loading json calendar files. Made changes to the
pom.xml file to support the new feature and made sure they communicate
by setting correct versions. Removed previous greeting example. Also
removed spring-boot dependencies in parent pom.

There are no breaking changes with this commit, but note that the api
module doesn't really communicate with the rest of the project yet.