chore: add unirest in javafx client
This commit adds basic unirest functionality to the javafx client.
It does include changes to pom.xml in fxui and core.
I've also added a HTTPCaller.java class in core which is for testing
and will be changed when the RestAPI is further along. In general,
this version is meant to show that unirest is working and can communicate
with the RestAPI. Tests have been commented out in FamControllerTest.java
as they don't pass.

There are breaking changes in this commit and the build is unusable
for most things except communication with the api.