fix: HTTPCaller now returns calendar from API
Added proper deserialiation of calendar from API in HTTPCaller.
Will move the deserialiation to CalendarDeserializer in the future.
Made changes to FamController to reflect the changes in HTTPCaller.
This means the JavaFX client can now pull the calendar from the API.

Breaking changes: Creating a new booking does not work.