feat: restapi now supports posting of bookings
This commit includes a lot of changes. The RestAPI now supports
posting of bookings. Currently this is done by sending a json-formatted
string. Will change this to be proper json later.
The JavaFX client has been updated to load the calendar from the API,
and to post bookings to the API. In addition it can delete bookings.
At the moment, the api is having trouble actually deleting bookings as
it is only happening clientside. Will investigate this later.

There are breaking changes! Tests fail and the delete button isn't really
working. Some tests have been commented out!