refactor: added json.simple source code to project
Due to json.simple being a dependency of the project, its source code
is now put into a new module named "json_simple". This is because
json.simple doesn't have a stable module name and therefore we cannot
package the project without integrating it into the source code.

The restapi does not currently boot with this change.